Break up components into different files
