+++ packages/desktop-client/src/components/budget/BudgetCategoriesV2.tsx
@@ -3,10 +3,8 @@ import React, {
   useMemo,
   useContext,
   type ComponentPropsWithoutRef,
-  useRef,
   useCallback,
 } from 'react';
-import { useFocusable, useFocusVisible, usePress } from 'react-aria';
 import {
   Column,
   Table,
@@ -21,12 +19,7 @@ import {
 import { Trans, useTranslation } from 'react-i18next';
 
 import { Button } from '@actual-app/components/button';
-import { SvgExpandArrow } from '@actual-app/components/icons/v0';
-import {
-  SvgArrowThinRight,
-  SvgCheveronDown,
-  SvgDotsHorizontalTriple,
-} from '@actual-app/components/icons/v1';
+import { SvgDotsHorizontalTriple } from '@actual-app/components/icons/v1';
 import { Input } from '@actual-app/components/input';
 import { Menu } from '@actual-app/components/menu';
 import { Popover } from '@actual-app/components/popover';
@@ -34,14 +27,9 @@ import { type CSSProperties } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
-import { css, cx } from '@emotion/css';
+import { css } from '@emotion/css';
 
 import * as monthUtils from 'loot-core/shared/months';
-import {
-  currencyToAmount,
-  currencyToInteger,
-  type IntegerAmount,
-} from 'loot-core/shared/util';
 import {
   type CategoryEntity,
   type CategoryGroupEntity,
@@ -51,10 +39,12 @@ import { BalanceMenu as EnvelopeBalanceMenu } from './envelope/BalanceMenu';
 import { BudgetMenu as EnvelopeBudgetMenu } from './envelope/BudgetMenu';
 import { CoverMenu } from './envelope/CoverMenu';
 import { TransferMenu } from './envelope/TransferMenu';
+import { ExpenseCategoryRow } from './ExpenseCategoryRow';
+import { ExpenseGroupRow } from './ExpenseGroupRow';
+import { IncomeCategoryRow } from './IncomeCategoryRow';
+import { IncomeGroupRow } from './IncomeGroupRow';
 import { MonthsContext } from './MonthsContext';
-import { BalanceMenu as TrackingBalanceMenu } from './tracking/BalanceMenu';
-import { BudgetMenu as TrackingBudgetMenu } from './tracking/BudgetMenu';
-import { makeAmountGrey, makeBalanceAmountStyle, separateGroups } from './util';
+import { separateGroups } from './util';
 
 import { NotesButton } from '@desktop-client/components/NotesButton';
 import {
@@ -76,11 +66,11 @@ import {
 
 const ROW_HEIGHT = 30;
 
-const balanceColumnPaddingStyle: CSSProperties = {
+export const balanceColumnPaddingStyle: CSSProperties = {
   paddingRight: 8,
 };
 
-const getCategoryRowStyle = (
+export const getCategoryRowStyle = (
   category: CategoryEntity | null = null,
   categoryGroup: CategoryGroupEntity | null = null,
 ): CSSProperties => ({
@@ -90,7 +80,7 @@ const getCategoryRowStyle = (
   opacity: category?.hidden || categoryGroup?.hidden ? 0.5 : 1,
 });
 
-const getCategoryGroupRowStyle = (
+export const getCategoryGroupRowStyle = (
   categoryGroup: CategoryGroupEntity | null = null,
 ): CSSProperties => ({
   height: ROW_HEIGHT,
@@ -111,7 +101,7 @@ const headerCellStyle: CSSProperties = {
   fontWeight: 600,
 };
 
-const hoverVisibleStyle: CSSProperties = {
+export const hoverVisibleStyle: CSSProperties = {
   '& .hover-visible': {
     opacity: 0,
   },
@@ -156,7 +146,7 @@ type Item =
       id: 'new-category';
     };
 
-type ColumnDefinition = {
+export type ColumnDefinition = {
   key: string;
   month: string;
   type: 'category' | 'budgeted' | 'spent' | 'balance';
@@ -387,10 +377,6 @@ export function BudgetCategories({
       style={{
         marginLeft: 4,
         marginRight: 4,
-        marginBottom: 10,
-        // Needed for border radius to work.
-        overflow: 'hidden',
-        borderRadius: 4,
       }}
       onKeyDown={e => {
         if (e.key === 'Escape' && e.target instanceof HTMLElement) {
@@ -410,6 +396,7 @@ export function BudgetCategories({
           aria-label={t('Budget table')}
           className={css({
             backgroundColor: theme.tableBackground,
+            marginBottom: 10,
             borderSpacing: 0,
             'th, td': {
               borderBottom: `1px solid ${theme.tableBorder}`,
@@ -782,417 +769,7 @@ function NewCategoryGroupRow({
   );
 }
 
-type CategorySpentCellProps = ComponentPropsWithoutRef<typeof ReactAriaCell> & {
-  month: string;
-  category: CategoryEntity;
-  onShowActivity: (category: CategoryEntity, month: string) => void;
-};
-
-function CategorySpentCell({
-  month,
-  category,
-  onShowActivity,
-  style,
-  ...props
-}: CategorySpentCellProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-
-  const categorySpentBinding =
-    budgetType === 'rollover'
-      ? envelopeBudget.catSumAmount(category.id)
-      : trackingBudget.catSumAmount(category.id);
-
-  const { pressProps } = usePress({
-    onPress: () => onShowActivity(category, month),
-  });
-
-  const textRef = useRef<HTMLSpanElement | null>(null);
-  const { focusableProps } = useFocusable(
-    {
-      onKeyUp: e => {
-        if (e.key === 'Enter') {
-          onShowActivity(category, month);
-        }
-      },
-    },
-    textRef,
-  );
-
-  return (
-    <ReactAriaCell style={{ textAlign: 'right', ...style }} {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <CellValue<'envelope-budget', 'sum-amount'>
-          type="financial"
-          binding={categorySpentBinding}
-        >
-          {props => (
-            <CellValueText
-              innerRef={textRef}
-              {...pressProps}
-              {...focusableProps}
-              {...props}
-              className={css({
-                ...makeAmountGrey(props.value),
-                '&:hover': {
-                  cursor: 'pointer',
-                  textDecoration: 'underline',
-                },
-                paddingRight: 5,
-              })}
-            />
-          )}
-        </CellValue>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-type CategoryBalanceCellProps = ComponentPropsWithoutRef<
-  typeof ReactAriaCell
-> & {
-  month: string;
-  category: CategoryEntity;
-  onBudgetAction: (month: string, action: string, args: unknown) => void;
-};
-
-function CategoryBalanceCell({
-  month,
-  category,
-  onBudgetAction,
-  style,
-  ...props
-}: CategoryBalanceCellProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  const triggerRef = useRef<HTMLSpanElement | null>(null);
-
-  const bindingBudgetType: SheetNames =
-    budgetType === 'rollover' ? 'envelope-budget' : 'tracking-budget';
-
-  const categoryCarryoverBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.catCarryover(category.id)
-      : trackingBudget.catCarryover(category.id);
-
-  const categoryBalanceBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.catBalance(category.id)
-      : trackingBudget.catBalance(category.id);
-
-  const categoryBudgetedBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.catBudgeted(category.id)
-      : trackingBudget.catBudgeted(category.id);
-
-  const categoryGoalBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.catGoal(category.id)
-      : trackingBudget.catGoal(category.id);
-
-  const categoryLongGoalBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.catLongGoal(category.id)
-      : trackingBudget.catLongGoal(category.id);
-
-  const budgetedValue = useSheetValue<
-    typeof bindingBudgetType,
-    typeof categoryBudgetedBinding
-  >(categoryBudgetedBinding);
-
-  const balanceValue = useSheetValue<
-    typeof bindingBudgetType,
-    typeof categoryBalanceBinding
-  >(categoryBalanceBinding);
-
-  const goalValue = useSheetValue<
-    typeof bindingBudgetType,
-    typeof categoryGoalBinding
-  >(categoryGoalBinding);
-
-  const longGoalValue = useSheetValue<
-    typeof bindingBudgetType,
-    typeof categoryLongGoalBinding
-  >(categoryLongGoalBinding);
-
-  const [activeBalanceMenu, setActiveBalanceMenu] = useState<
-    'balance' | 'transfer' | 'cover' | null
-  >(null);
-
-  const { pressProps } = usePress({
-    onPress: () => setActiveBalanceMenu('balance'),
-  });
-
-  const { focusableProps } = useFocusable(
-    {
-      onKeyUp: e => {
-        if (e.key === 'Enter') {
-          setActiveBalanceMenu('balance');
-        }
-      },
-    },
-    triggerRef,
-  );
-
-  const isGoalTemplatesEnabled = useFeatureFlag('goalTemplatesEnabled');
-
-  const getBalanceAmountStyle = useCallback(
-    (balanceValue: number) =>
-      makeBalanceAmountStyle(
-        balanceValue,
-        isGoalTemplatesEnabled ? goalValue : null,
-        longGoalValue === 1 ? balanceValue : budgetedValue,
-      ),
-    [budgetedValue, goalValue, isGoalTemplatesEnabled, longGoalValue],
-  );
-
-  // TODO: Refactor balance cell tooltips
-  return (
-    <ReactAriaCell style={{ textAlign: 'right', ...style }} {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <CellValue<typeof bindingBudgetType, typeof categoryBalanceBinding>
-          type="financial"
-          binding={categoryBalanceBinding}
-        >
-          {balanceProps => (
-            <View
-              style={{
-                position: 'relative',
-                display: 'inline-block',
-                ...balanceColumnPaddingStyle,
-              }}
-            >
-              <CellValueText
-                innerRef={triggerRef}
-                {...pressProps}
-                {...focusableProps}
-                {...balanceProps}
-                className={css({
-                  ...getBalanceAmountStyle(balanceProps.value),
-                  '&:hover': {
-                    cursor: 'pointer',
-                    textDecoration: 'underline',
-                  },
-                })}
-              />
-              <CellValue<
-                typeof bindingBudgetType,
-                typeof categoryCarryoverBinding
-              >
-                binding={categoryCarryoverBinding}
-              >
-                {carryOverProps =>
-                  carryOverProps.value && (
-                    <CarryoverIndicator
-                      style={getBalanceAmountStyle(balanceProps.value)}
-                    />
-                  )
-                }
-              </CellValue>
-            </View>
-          )}
-        </CellValue>
-        <Popover
-          triggerRef={triggerRef}
-          placement="bottom end"
-          isOpen={activeBalanceMenu !== null}
-          onOpenChange={() => {
-            if (activeBalanceMenu !== 'balance') {
-              setActiveBalanceMenu(null);
-            }
-          }}
-          isNonModal
-        >
-          {budgetType === 'rollover' ? (
-            <>
-              {activeBalanceMenu === 'balance' && (
-                <EnvelopeBalanceMenu
-                  categoryId={category.id}
-                  onCarryover={carryover => {
-                    onBudgetAction(month, 'carryover', {
-                      category: category.id,
-                      flag: carryover,
-                    });
-                    setActiveBalanceMenu(null);
-                  }}
-                  onTransfer={() => setActiveBalanceMenu('transfer')}
-                  onCover={() => setActiveBalanceMenu('cover')}
-                />
-              )}
-              {activeBalanceMenu === 'transfer' && (
-                <TransferMenu
-                  categoryId={category.id}
-                  initialAmount={balanceValue}
-                  showToBeBudgeted={true}
-                  onSubmit={(amount, toCategoryId) => {
-                    onBudgetAction(month, 'transfer-category', {
-                      amount,
-                      from: category.id,
-                      to: toCategoryId,
-                    });
-                  }}
-                  onClose={() => setActiveBalanceMenu(null)}
-                />
-              )}
-              {activeBalanceMenu === 'cover' && (
-                <CoverMenu
-                  categoryId={category.id}
-                  onSubmit={fromCategoryId => {
-                    onBudgetAction(month, 'cover-overspending', {
-                      to: category.id,
-                      from: fromCategoryId,
-                    });
-                  }}
-                  onClose={() => setActiveBalanceMenu(null)}
-                />
-              )}
-            </>
-          ) : (
-            <TrackingBalanceMenu
-              categoryId={category.id}
-              onCarryover={carryover => {
-                onBudgetAction(month, 'carryover', {
-                  category: category.id,
-                  flag: carryover,
-                });
-                setActiveBalanceMenu(null);
-              }}
-            />
-          )}
-        </Popover>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-type CategoryGroupBudgetedCellProps = ComponentPropsWithoutRef<
-  typeof ReactAriaCell
-> & {
-  month: string;
-  categoryGroup: CategoryGroupEntity;
-};
-
-function CategoryGroupBudgetedCell({
-  month,
-  categoryGroup,
-  style,
-  ...props
-}: CategoryGroupBudgetedCellProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  const bindingBudgetType: SheetNames =
-    budgetType === 'rollover' ? 'envelope-budget' : 'tracking-budget';
-
-  const groupBudgetedBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.groupBudgeted(categoryGroup.id)
-      : trackingBudget.groupBudgeted(categoryGroup.id);
-
-  return (
-    <ReactAriaCell style={{ textAlign: 'right', ...style }} {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <CellValue<typeof bindingBudgetType, typeof groupBudgetedBinding>
-          type="financial"
-          binding={groupBudgetedBinding}
-        >
-          {props => (
-            <CellValueText
-              {...props}
-              style={{
-                paddingRight: 5,
-              }}
-            />
-          )}
-        </CellValue>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-type CategoryGroupSpentCellProps = ComponentPropsWithoutRef<
-  typeof ReactAriaCell
-> & {
-  month: string;
-  categoryGroup: CategoryGroupEntity;
-};
-
-function CategoryGroupSpentCell({
-  month,
-  categoryGroup,
-  style,
-  ...props
-}: CategoryGroupSpentCellProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  const bindingBudgetType: SheetNames =
-    budgetType === 'rollover' ? 'envelope-budget' : 'tracking-budget';
-
-  const groupSpentBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.groupSumAmount(categoryGroup.id)
-      : trackingBudget.groupSumAmount(categoryGroup.id);
-
-  return (
-    <ReactAriaCell style={{ textAlign: 'right', ...style }} {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <CellValue<typeof bindingBudgetType, typeof groupSpentBinding>
-          type="financial"
-          binding={groupSpentBinding}
-        >
-          {props => (
-            <CellValueText
-              {...props}
-              style={{
-                paddingRight: 5,
-              }}
-            />
-          )}
-        </CellValue>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-type CategoryGroupBalanceCellProps = ComponentPropsWithoutRef<
-  typeof ReactAriaCell
-> & {
-  month: string;
-  categoryGroup: CategoryGroupEntity;
-};
-
-function CategoryGroupBalanceCell({
-  month,
-  categoryGroup,
-  style,
-  ...props
-}: CategoryGroupBalanceCellProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  const bindingBudgetType: SheetNames =
-    budgetType === 'rollover' ? 'envelope-budget' : 'tracking-budget';
-
-  const groupBalanceBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.groupBalance(categoryGroup.id)
-      : trackingBudget.groupBalance(categoryGroup.id);
-
-  return (
-    <ReactAriaCell style={{ textAlign: 'right', ...style }} {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <CellValue<typeof bindingBudgetType, typeof groupBalanceBinding>
-          type="financial"
-          binding={groupBalanceBinding}
-        >
-          {props => (
-            <CellValueText
-              {...props}
-              style={{
-                ...balanceColumnPaddingStyle,
-              }}
-            />
-          )}
-        </CellValue>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-const getHeaderBackgroundStyle = (
+export const getHeaderBackgroundStyle = (
   type: 'category' | 'budgeted' | 'spent' | 'balance',
   month: string,
 ): CSSProperties => {
@@ -1205,92 +782,7 @@ const getHeaderBackgroundStyle = (
   return { backgroundColor: theme.budgetHeaderOtherMonth };
 };
 
-type ExpenseGroupRowProps = ComponentPropsWithoutRef<
-  typeof ReactAriaRow<ColumnDefinition>
-> & {
-  item: {
-    type: 'expense-group';
-    id: `expense-group-${string}`;
-    value: CategoryGroupEntity;
-  };
-  isCollapsed: boolean;
-  onToggleCollapse: (categoryGroup: CategoryGroupEntity) => void;
-  onAddCategory: (categoryGroup: CategoryGroupEntity) => void;
-  onRename: (categoryGroup: CategoryGroupEntity, newName: string) => void;
-  onDelete: (categoryGroup: CategoryGroupEntity) => void;
-  onToggleVisibilty: (categoryGroup: CategoryGroupEntity) => void;
-  onApplyBudgetTemplatesInGroup: (categoryGroup: CategoryGroupEntity) => void;
-};
-
-function ExpenseGroupRow({
-  item,
-  isCollapsed,
-  onToggleCollapse,
-  onAddCategory,
-  onRename,
-  onDelete,
-  onToggleVisibilty,
-  onApplyBudgetTemplatesInGroup,
-  style,
-  ...props
-}: ExpenseGroupRowProps) {
-  return (
-    <ReactAriaRow
-      style={{
-        ...getCategoryGroupRowStyle(item.value),
-        ...style,
-      }}
-      {...props}
-    >
-      {column => {
-        switch (column.type) {
-          case 'category':
-            return (
-              <CategoryGroupNameCell
-                month={column.month}
-                categoryGroup={item.value}
-                isCollapsed={isCollapsed}
-                onToggleCollapse={onToggleCollapse}
-                onAddCategory={onAddCategory}
-                onRename={onRename}
-                onDelete={onDelete}
-                onToggleVisibilty={onToggleVisibilty}
-                onApplyBudgetTemplatesInGroup={onApplyBudgetTemplatesInGroup}
-              />
-            );
-          case 'budgeted':
-            return (
-              <CategoryGroupBudgetedCell
-                month={column.month}
-                categoryGroup={item.value}
-                style={getHeaderBackgroundStyle(column.type, column.month)}
-              />
-            );
-          case 'spent':
-            return (
-              <CategoryGroupSpentCell
-                month={column.month}
-                categoryGroup={item.value}
-                style={getHeaderBackgroundStyle(column.type, column.month)}
-              />
-            );
-          case 'balance':
-            return (
-              <CategoryGroupBalanceCell
-                month={column.month}
-                categoryGroup={item.value}
-                style={getHeaderBackgroundStyle(column.type, column.month)}
-              />
-            );
-          default:
-            throw new Error(`Unrecognized column type: ${column.type}`);
-        }
-      }}
-    </ReactAriaRow>
-  );
-}
-
-const getCellBackgroundStyle = (
+export const getCellBackgroundStyle = (
   type: 'category' | 'budgeted' | 'spent' | 'balance',
   month: string,
 ): CSSProperties => {
@@ -1302,783 +794,6 @@ const getCellBackgroundStyle = (
   }
   return { backgroundColor: theme.budgetOtherMonth };
 };
-
-type ExpenseCategoryRowProps = ComponentPropsWithoutRef<
-  typeof ReactAriaRow<ColumnDefinition>
-> & {
-  item: {
-    type: 'expense-category';
-    id: `expense-category-${string}`;
-    value: CategoryEntity;
-    group: CategoryGroupEntity;
-  };
-  onBudgetAction: (month: string, action: string, args: unknown) => void;
-  onShowActivity: (category: CategoryEntity, month: string) => void;
-  onRename: (category: CategoryEntity, newName: string) => void;
-  onDelete: (category: CategoryEntity) => void;
-  onToggleVisibility: (category: CategoryEntity) => void;
-};
-
-function ExpenseCategoryRow({
-  item,
-  onBudgetAction,
-  onShowActivity,
-  onRename,
-  onDelete,
-  onToggleVisibility,
-  style,
-  ...props
-}: ExpenseCategoryRowProps) {
-  return (
-    <ReactAriaRow
-      style={{
-        ...getCategoryRowStyle(item.value, item.group),
-        ...style,
-      }}
-      {...props}
-    >
-      {column => {
-        switch (column.type) {
-          case 'category':
-            return (
-              <CategoryNameCell
-                month={column.month}
-                category={item.value}
-                categoryGroup={item.group}
-                onRename={onRename}
-                onDelete={onDelete}
-                onToggleVisibility={onToggleVisibility}
-              />
-            );
-          case 'budgeted':
-            return (
-              <CategoryBudgetedCell
-                month={column.month}
-                category={item.value}
-                onBudgetAction={onBudgetAction}
-                style={getCellBackgroundStyle(column.type, column.month)}
-              />
-            );
-          case 'spent':
-            return (
-              <CategorySpentCell
-                month={column.month}
-                category={item.value}
-                onShowActivity={onShowActivity}
-                style={getCellBackgroundStyle(column.type, column.month)}
-              />
-            );
-          case 'balance':
-            return (
-              <CategoryBalanceCell
-                month={column.month}
-                category={item.value}
-                onBudgetAction={onBudgetAction}
-                style={getCellBackgroundStyle(column.type, column.month)}
-              />
-            );
-          default:
-            throw new Error(`Unrecognized column type: ${column.type}`);
-        }
-      }}
-    </ReactAriaRow>
-  );
-}
-
-type IncomeGroupRowProps = ComponentPropsWithoutRef<
-  typeof ReactAriaRow<ColumnDefinition>
-> & {
-  item: {
-    type: 'income-group';
-    id: `income-group-${string}`;
-    value: CategoryGroupEntity;
-  };
-  isCollapsed: boolean;
-  onToggleCollapse: (categoryGroup: CategoryGroupEntity) => void;
-  onAddCategory: (categoryGroup: CategoryGroupEntity) => void;
-  onRename: (categoryGroup: CategoryGroupEntity, newName: string) => void;
-  onDelete: (categoryGroup: CategoryGroupEntity) => void;
-  onToggleVisibilty: (categoryGroup: CategoryGroupEntity) => void;
-  onApplyBudgetTemplatesInGroup: (categoryGroup: CategoryGroupEntity) => void;
-};
-
-function IncomeGroupRow({
-  item,
-  isCollapsed,
-  onToggleCollapse,
-  onAddCategory,
-  onRename,
-  onDelete,
-  onToggleVisibilty,
-  onApplyBudgetTemplatesInGroup,
-  style,
-  ...props
-}: IncomeGroupRowProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-
-  return budgetType === 'rollover' ? (
-    <ReactAriaRow
-      style={{
-        ...getCategoryGroupRowStyle(item.value),
-        ...style,
-      }}
-      {...props}
-    >
-      {column => {
-        switch (column.type) {
-          case 'category':
-            return (
-              <CategoryGroupNameCell
-                month={column.month}
-                categoryGroup={item.value}
-                isCollapsed={isCollapsed}
-                onToggleCollapse={onToggleCollapse}
-                onAddCategory={onAddCategory}
-                onRename={onRename}
-                onDelete={onDelete}
-                onToggleVisibilty={onToggleVisibilty}
-                onApplyBudgetTemplatesInGroup={onApplyBudgetTemplatesInGroup}
-              />
-            );
-          case 'budgeted':
-            return <ReactAriaCell />;
-          case 'spent':
-            return <ReactAriaCell />;
-          case 'balance':
-            return (
-              <CategoryGroupBalanceCell
-                month={column.month}
-                categoryGroup={item.value}
-              />
-            );
-          default:
-            throw new Error(`Unrecognized column type: ${column.type}`);
-        }
-      }}
-    </ReactAriaRow>
-  ) : (
-    <ReactAriaRow
-      style={{
-        ...getCategoryGroupRowStyle(item.value),
-        ...style,
-      }}
-      {...props}
-    >
-      {column => {
-        switch (column.type) {
-          case 'category':
-            return (
-              <CategoryGroupNameCell
-                month={column.month}
-                categoryGroup={item.value}
-                isCollapsed={isCollapsed}
-                onToggleCollapse={onToggleCollapse}
-                onAddCategory={onAddCategory}
-                onRename={onRename}
-                onDelete={onDelete}
-                onToggleVisibilty={onToggleVisibilty}
-                onApplyBudgetTemplatesInGroup={onApplyBudgetTemplatesInGroup}
-              />
-            );
-          case 'budgeted':
-            return (
-              <CategoryGroupBudgetedCell
-                month={column.month}
-                categoryGroup={item.value}
-              />
-            );
-          case 'spent':
-            return <ReactAriaCell />;
-          case 'balance':
-            return (
-              <CategoryGroupBalanceCell
-                month={column.month}
-                categoryGroup={item.value}
-              />
-            );
-          default:
-            throw new Error(`Unrecognized column type: ${column.type}`);
-        }
-      }}
-    </ReactAriaRow>
-  );
-}
-
-type IncomeCategoryRowProps = ComponentPropsWithoutRef<
-  typeof ReactAriaRow<ColumnDefinition>
-> & {
-  item: {
-    type: 'income-category';
-    id: `income-category-${string}`;
-    value: CategoryEntity;
-    group: CategoryGroupEntity;
-  };
-  onBudgetAction: (month: string, action: string, args: unknown) => void;
-  onRename: (category: CategoryEntity, newName: string) => void;
-  onDelete: (category: CategoryEntity) => void;
-  onToggleVisibility: (category: CategoryEntity) => void;
-};
-
-function IncomeCategoryRow({
-  item,
-  onBudgetAction,
-  onRename,
-  onDelete,
-  onToggleVisibility,
-  style,
-  ...props
-}: IncomeCategoryRowProps) {
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  return budgetType === 'rollover' ? (
-    <ReactAriaRow
-      style={{
-        ...getCategoryRowStyle(item.value),
-        ...style,
-      }}
-      {...props}
-    >
-      {column => {
-        switch (column.type) {
-          case 'category':
-            return (
-              <CategoryNameCell
-                month={column.month}
-                category={item.value}
-                categoryGroup={item.group}
-                onRename={onRename}
-                onDelete={onDelete}
-                onToggleVisibility={onToggleVisibility}
-              />
-            );
-          case 'budgeted':
-            return <ReactAriaCell />;
-          case 'spent':
-            return <ReactAriaCell />;
-          case 'balance':
-            return (
-              <CategoryBalanceCell
-                month={column.month}
-                category={item.value}
-                onBudgetAction={onBudgetAction}
-              />
-            );
-          default:
-            throw new Error(`Unrecognized column type: ${column.type}`);
-        }
-      }}
-    </ReactAriaRow>
-  ) : (
-    <ReactAriaRow
-      style={{
-        ...getCategoryRowStyle(item.value),
-        ...style,
-      }}
-      {...props}
-    >
-      {column => {
-        switch (column.type) {
-          case 'category':
-            return (
-              <CategoryNameCell
-                month={column.month}
-                category={item.value}
-                categoryGroup={item.group}
-                onRename={onRename}
-                onDelete={onDelete}
-                onToggleVisibility={onToggleVisibility}
-              />
-            );
-          case 'budgeted':
-            return (
-              <CategoryBudgetedCell
-                month={column.month}
-                category={item.value}
-                onBudgetAction={onBudgetAction}
-              />
-            );
-          case 'spent':
-            return <ReactAriaCell />;
-          case 'balance':
-            return (
-              <CategoryBalanceCell
-                month={column.month}
-                category={item.value}
-                onBudgetAction={onBudgetAction}
-              />
-            );
-          default:
-            throw new Error(`Unrecognized column type: ${column.type}`);
-        }
-      }}
-    </ReactAriaRow>
-  );
-}
-
-type BudgetedInputProps = Omit<
-  ComponentPropsWithoutRef<typeof Input>,
-  'value'
-> & {
-  value: IntegerAmount;
-  onUpdateAmount: (newValue: IntegerAmount) => void;
-};
-
-function BudgetedInput({
-  value,
-  onFocus,
-  onChangeValue,
-  onUpdate,
-  onUpdateAmount,
-  ...props
-}: BudgetedInputProps) {
-  const format = useFormat();
-  const [currentFormattedAmount, setCurrentFormattedAmount] = useState<
-    string | null
-  >(null);
-
-  return (
-    <Input
-      value={currentFormattedAmount ?? format(value, 'financial')}
-      onFocus={e => {
-        onFocus?.(e);
-        if (!e.defaultPrevented) {
-          e.target.select();
-        }
-      }}
-      onEscape={() => setCurrentFormattedAmount(format(value, 'financial'))}
-      className={css({
-        ...makeAmountGrey(
-          currentFormattedAmount
-            ? currencyToAmount(currentFormattedAmount)
-            : value,
-        ),
-        textAlign: 'right',
-        border: '1px solid transparent',
-        '&:hover:not(:focus)': {
-          border: `1px solid ${theme.formInputBorder}`,
-        },
-      })}
-      onChangeValue={(newValue, e) => {
-        onChangeValue?.(newValue, e);
-        setCurrentFormattedAmount(newValue);
-      }}
-      onUpdate={(newValue, e) => {
-        onUpdate?.(newValue, e);
-        const integerAmount = currencyToInteger(newValue);
-        onUpdateAmount?.(integerAmount);
-        setCurrentFormattedAmount(format(integerAmount, 'financial'));
-      }}
-      {...props}
-    />
-  );
-}
-
-type CategoryBudgetedCellProps = ComponentPropsWithoutRef<
-  typeof ReactAriaCell
-> & {
-  month: string;
-  category: CategoryEntity;
-  onBudgetAction: (month: string, action: string, args: unknown) => void;
-};
-
-function CategoryBudgetedCell({
-  month,
-  category,
-  onBudgetAction,
-  ...props
-}: CategoryBudgetedCellProps) {
-  const { t } = useTranslation();
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [shouldHideBudgetMenuButton, setShouldHideBudgetMenuButton] =
-    useState(false);
-
-  const bindingBudgetType: SheetNames =
-    budgetType === 'rollover' ? 'envelope-budget' : 'tracking-budget';
-
-  const budgetedBinding =
-    bindingBudgetType === 'envelope-budget'
-      ? envelopeBudget.catBudgeted(category.id)
-      : trackingBudget.catBudgeted(category.id);
-
-  const BudgetMenuComponent =
-    bindingBudgetType === 'envelope-budget'
-      ? EnvelopeBudgetMenu
-      : TrackingBudgetMenu;
-
-  const { showUndoNotification } = useUndo();
-
-  const onUpdateBudget = (amount: IntegerAmount) => {
-    onBudgetAction(month, 'budget-amount', {
-      category: category.id,
-      amount,
-    });
-  };
-
-  const { isFocusVisible } = useFocusVisible();
-
-  return (
-    <ReactAriaCell {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <View
-          className={css({
-            flexDirection: 'row',
-            alignItems: 'center',
-            ...hoverVisibleStyle,
-          })}
-        >
-          <DialogTrigger>
-            <Button
-              variant="bare"
-              aria-label={t('Budget menu')}
-              className={cx(
-                { 'hover-visible': !isMenuOpen && !isFocusVisible },
-                css({
-                  display:
-                    shouldHideBudgetMenuButton && !isFocusVisible
-                      ? 'none'
-                      : undefined,
-                }),
-              )}
-              onPress={() => setIsMenuOpen(true)}
-            >
-              <SvgCheveronDown width={12} height={12} />
-            </Button>
-
-            <Popover
-              placement="bottom start"
-              isOpen={isMenuOpen}
-              onOpenChange={() => setIsMenuOpen(false)}
-              isNonModal
-            >
-              <BudgetMenuComponent
-                onCopyLastMonthAverage={() => {
-                  onBudgetAction(month, 'copy-single-last', {
-                    category: category.id,
-                  });
-                  showUndoNotification({
-                    message: t(`Budget set to last month’s budget.`),
-                  });
-                  setIsMenuOpen(false);
-                }}
-                onSetMonthsAverage={numberOfMonths => {
-                  if (
-                    numberOfMonths !== 3 &&
-                    numberOfMonths !== 6 &&
-                    numberOfMonths !== 12
-                  ) {
-                    return;
-                  }
-                  onBudgetAction(month, `set-single-${numberOfMonths}-avg`, {
-                    category: category.id,
-                  });
-                  showUndoNotification({
-                    message: t(
-                      'Budget set to {{numberOfMonths}}-month average.',
-                      { numberOfMonths },
-                    ),
-                  });
-                  setIsMenuOpen(false);
-                }}
-                onApplyBudgetTemplate={() => {
-                  onBudgetAction(month, 'apply-single-category-template', {
-                    category: category.id,
-                  });
-                  showUndoNotification({
-                    message: t(`Budget template applied.`),
-                  });
-                  setIsMenuOpen(false);
-                }}
-              />
-            </Popover>
-          </DialogTrigger>
-          <View style={{ flex: 1 }}>
-            <CellValue<typeof bindingBudgetType, typeof budgetedBinding>
-              type="financial"
-              binding={budgetedBinding}
-            >
-              {({ value: budgetedAmount }) => (
-                <BudgetedInput
-                  value={budgetedAmount}
-                  onFocus={() => setShouldHideBudgetMenuButton(true)}
-                  onBlur={() => setShouldHideBudgetMenuButton(false)}
-                  style={getCellBackgroundStyle('budgeted', month)}
-                  onUpdateAmount={onUpdateBudget}
-                />
-              )}
-            </CellValue>
-          </View>
-        </View>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-type CategoryGroupNameCellProps = ComponentPropsWithoutRef<
-  typeof ReactAriaCell
-> & {
-  month: string;
-  categoryGroup: CategoryGroupEntity;
-  isCollapsed: boolean;
-  onToggleCollapse: (categoryGroup: CategoryGroupEntity) => void;
-  onAddCategory: (categoryGroup: CategoryGroupEntity) => void;
-  onRename: (categoryGroup: CategoryGroupEntity, newName: string) => void;
-  onDelete: (categoryGroup: CategoryGroupEntity) => void;
-  onToggleVisibilty: (categoryGroup: CategoryGroupEntity) => void;
-  onApplyBudgetTemplatesInGroup: (categoryGroup: CategoryGroupEntity) => void;
-};
-
-function CategoryGroupNameCell({
-  month,
-  categoryGroup,
-  isCollapsed,
-  onToggleCollapse,
-  onAddCategory,
-  onRename,
-  onDelete,
-  onToggleVisibilty,
-  onApplyBudgetTemplatesInGroup,
-  ...props
-}: CategoryGroupNameCellProps) {
-  const { t } = useTranslation();
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [isRenaming, setIsRenaming] = useState(false);
-  const isGoalTemplatesEnabled = useFeatureFlag('goalTemplatesEnabled');
-  const { isFocusVisible } = useFocusVisible();
-
-  return (
-    <ReactAriaCell {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <View
-          className={css({
-            paddingLeft: 5,
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-            ...hoverVisibleStyle,
-          })}
-        >
-          <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-            <Button
-              variant="bare"
-              onPress={() => onToggleCollapse(categoryGroup)}
-              isDisabled={isRenaming}
-            >
-              <SvgExpandArrow
-                width={8}
-                height={8}
-                style={{
-                  flexShrink: 0,
-                  transition: 'transform .1s',
-                  transform: isCollapsed ? 'rotate(-90deg)' : '',
-                }}
-              />
-            </Button>
-            {isRenaming ? (
-              <View style={{ flex: 1 }}>
-                <Input
-                  autoFocus
-                  defaultValue={categoryGroup.name}
-                  onBlur={() => setIsRenaming(false)}
-                  onEscape={() => setIsRenaming(false)}
-                  onUpdate={newName => {
-                    if (newName !== categoryGroup.name) {
-                      onRename(categoryGroup, newName);
-                    }
-                  }}
-                />
-              </View>
-            ) : (
-              <>
-                <Text style={{ fontWeight: 600 }}>{categoryGroup.name}</Text>
-                <DialogTrigger>
-                  <Button
-                    variant="bare"
-                    className={cx(
-                      { 'hover-visible': !isMenuOpen && !isFocusVisible },
-                      css({ marginLeft: 5 }),
-                    )}
-                    onPress={() => {
-                      // resetPosition();
-                      setIsMenuOpen(true);
-                    }}
-                  >
-                    <SvgCheveronDown width={12} height={12} />
-                  </Button>
-
-                  <Popover
-                    placement="bottom start"
-                    isOpen={isMenuOpen}
-                    onOpenChange={() => setIsMenuOpen(false)}
-                    isNonModal
-                  >
-                    <Menu
-                      onMenuSelect={type => {
-                        if (type === 'rename') {
-                          // onEdit(categoryGroup.id);
-                          setIsRenaming(true);
-                        } else if (type === 'add-category') {
-                          onAddCategory(categoryGroup);
-                        } else if (type === 'delete') {
-                          onDelete(categoryGroup);
-                        } else if (type === 'toggle-visibility') {
-                          // onSave({ ...categoryGroup, hidden: !categoryGroup.hidden });
-                          onToggleVisibilty(categoryGroup);
-                        } else if (
-                          type === 'apply-multiple-category-template'
-                        ) {
-                          onApplyBudgetTemplatesInGroup(categoryGroup);
-                        }
-                        setIsMenuOpen(false);
-                      }}
-                      items={[
-                        { name: 'add-category', text: t('Add category') },
-                        { name: 'rename', text: t('Rename') },
-                        !categoryGroup.is_income && {
-                          name: 'toggle-visibility',
-                          text: categoryGroup.hidden ? 'Show' : 'Hide',
-                        },
-                        // onDelete && { name: 'delete', text: t('Delete') },
-                        { name: 'delete', text: t('Delete') },
-                        ...(isGoalTemplatesEnabled
-                          ? [
-                              {
-                                name: 'apply-multiple-category-template',
-                                text: t('Apply budget templates'),
-                              },
-                            ]
-                          : []),
-                      ]}
-                    />
-                  </Popover>
-                </DialogTrigger>
-              </>
-            )}
-          </View>
-          {!isRenaming && (
-            <View>
-              <NotesButton
-                id={categoryGroup.id}
-                defaultColor={theme.pageTextLight}
-              />
-            </View>
-          )}
-        </View>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
-
-type CategoryNameCellProps = ComponentPropsWithoutRef<typeof ReactAriaCell> & {
-  month: string;
-  category: CategoryEntity;
-  categoryGroup: CategoryGroupEntity;
-  onRename: (category: CategoryEntity, newName: string) => void;
-  onDelete: (category: CategoryEntity) => void;
-  onToggleVisibility: (category: CategoryEntity) => void;
-};
-
-function CategoryNameCell({
-  month,
-  category,
-  categoryGroup,
-  onRename,
-  onDelete,
-  onToggleVisibility,
-  ...props
-}: CategoryNameCellProps) {
-  const { t } = useTranslation();
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [isRenaming, setIsRenaming] = useState(false);
-  const { isFocusVisible } = useFocusVisible();
-
-  return (
-    <ReactAriaCell {...props}>
-      <SheetNameProvider name={monthUtils.sheetForMonth(month)}>
-        <View
-          className={css({
-            paddingLeft: 18,
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-            ...hoverVisibleStyle,
-          })}
-        >
-          {isRenaming ? (
-            <View style={{ flex: 1 }}>
-              <Input
-                defaultValue={category.name}
-                placeholder={t('Enter category name')}
-                onBlur={() => setIsRenaming(false)}
-                onUpdate={newName => {
-                  if (newName !== category.name) {
-                    onRename(category, newName);
-                  }
-                }}
-              />
-            </View>
-          ) : (
-            <>
-              <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                <Text>{category.name}</Text>
-                <DialogTrigger>
-                  <Button
-                    variant="bare"
-                    className={cx(
-                      { 'hover-visible': !isMenuOpen && !isFocusVisible },
-                      css({ marginLeft: 5 }),
-                    )}
-                    onPress={() => {
-                      // resetPosition();
-                      setIsMenuOpen(true);
-                    }}
-                  >
-                    <SvgCheveronDown width={12} height={12} />
-                  </Button>
-
-                  <Popover
-                    placement="bottom start"
-                    isOpen={isMenuOpen}
-                    onOpenChange={() => setIsMenuOpen(false)}
-                    isNonModal
-                  >
-                    <Menu
-                      onMenuSelect={type => {
-                        if (type === 'rename') {
-                          // onEditName(category.id);
-                          setIsRenaming(true);
-                        } else if (type === 'delete') {
-                          onDelete(category);
-                        } else if (type === 'toggle-visibility') {
-                          // onSave({ ...category, hidden: !category.hidden });
-                          onToggleVisibility(category);
-                        }
-                        setIsMenuOpen(false);
-                      }}
-                      items={[
-                        { name: 'rename', text: t('Rename') },
-                        !categoryGroup?.hidden && {
-                          name: 'toggle-visibility',
-                          text: category.hidden ? t('Show') : t('Hide'),
-                        },
-                        { name: 'delete', text: t('Delete') },
-                      ]}
-                    />
-                  </Popover>
-                </DialogTrigger>
-              </View>
-              <View>
-                <NotesButton
-                  id={category.id}
-                  defaultColor={theme.pageTextLight}
-                />
-              </View>
-            </>
-          )}
-        </View>
-      </SheetNameProvider>
-    </ReactAriaCell>
-  );
-}
 
 type CategoryColumnProps = ComponentPropsWithoutRef<typeof ResizableColumn> & {
   month: string;
@@ -2312,29 +1027,5 @@ function ResizableColumn({ children, ...props }: ResizableColumnProps) {
         </View>
       )}
     </Column>
-  );
-}
-
-type CarryoverIndicatorProps = {
-  style?: CSSProperties;
-};
-
-function CarryoverIndicator({ style }: CarryoverIndicatorProps) {
-  return (
-    <View
-      style={{
-        position: 'absolute',
-        right: 0,
-        transform: 'translateY(-50%)',
-        top: '50%',
-        ...style,
-      }}
-    >
-      <SvgArrowThinRight
-        width={style?.width || 7}
-        height={style?.height || 7}
-        style={style}
-      />
-    </View>
   );
 }
